Fix in help dialog
When using a text prompt to send the 2nd message in Help dialog, the bot will wait for user reply to that prompt instead of using the luis recognizer and listen to an user intent. This is because the botbuilder push a prompt "dialog" to the dialogstack and in the onEnabled event the logic is enabling or not the LuisRecognizer by checking the dialogStack length.